Added in my IK Scene & Dragging the hands with the mouse
Added in my IK rig scene as well as the art. I also added some code to clamp the collider on the hand, so it doesn't move away from the hand, allowing the player to keep dragging the hands. It detects the hands using a Raycast from the mouse and then moves it with the mouse.