Added an IK & Ragdoll hybrid arm
The limbs will be a ragdoll when you haven't clicked on them, then it will turn into an IK when you click on the hand. The hand collider is sometimes still in the wrong place and the IK wants to always snap the arm back to the right which breaks it.