Improved Camera Code
Used Cinemachine to help move the camera to where the players are. It also no gives the player their own camera if they move too far away from everyone else. It has some zoom based on the furthest away as well. It still needs testing in multiplayer and will require lots of tweaking and smoothing.