Fixed the breaker holds
Fixed the breaker holds throwing the player off when they spin. Moved the collider and script onto the parent and kept the animator and sprite on the child so the animation will now not affect the player before it has broken. Also added a check to see if we are in multiplayer before displaying the camera target group is null message.